executor: migrate test-infra to testify for testSerialSuite
