+++ executor/executor_legacy_test.go
@@ -17,9 +17,6 @@ package executor_test
 import (
 	"context"
 	"fmt"
-	"math"
-	"reflect"
-	"runtime"
 	"strconv"
 	"strings"
 	"testing"
@@ -30,9 +27,7 @@ import (
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/domain"
 	"github.com/pingcap/tidb/domain/infosync"
-	"github.com/pingcap/tidb/errno"
 	"github.com/pingcap/tidb/executor"
-	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/parser"
@@ -49,10 +44,8 @@ import (
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/admin"
-	"github.com/pingcap/tidb/util/memory"
 	"github.com/pingcap/tidb/util/mock"
 	"github.com/pingcap/tidb/util/testkit"
-	"github.com/pingcap/tidb/util/testleak"
 	"github.com/tikv/client-go/v2/oracle"
 	"github.com/tikv/client-go/v2/testutils"
 )
@@ -68,11 +61,9 @@ var _ = Suite(&testSuite{&baseTestSuite{}})
 var _ = Suite(&testSuiteP2{&baseTestSuite{}})
 var _ = Suite(&testSuite2{&baseTestSuite{}})
 var _ = Suite(&testSuite3{&baseTestSuite{}})
-var _ = SerialSuites(&testSerialSuite{&baseTestSuite{}})
 
 type testSuite struct{ *baseTestSuite }
 type testSuiteP2 struct{ *baseTestSuite }
-type testSerialSuite struct{ *baseTestSuite }
 
 type baseTestSuite struct {
 	cluster testutils.Cluster
@@ -2144,43 +2135,6 @@ func (s *testSuiteP2) TestApplyCache(c *C) {
 	c.Assert(value[ind:], Equals, "cache:OFF")
 }
 
-// For issue 17256
-func (s *testSuite) TestGenerateColumnReplace(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test;")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t1 (a int, b int as (a + 1) virtual not null, unique index idx(b));")
-	tk.MustExec("REPLACE INTO `t1` (`a`) VALUES (2);")
-	tk.MustExec("REPLACE INTO `t1` (`a`) VALUES (2);")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("2 3"))
-	tk.MustExec("insert into `t1` (`a`) VALUES (2) on duplicate key update a = 3;")
-	tk.MustQuery("select * from t1").Check(testkit.Rows("3 4"))
-}
-
-func (s *testSerialSuite) TestPrevStmtDesensitization(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test;")
-	tk.MustExec(fmt.Sprintf("set @@session.%v=1", variable.TiDBRedactLog))
-	defer tk.MustExec(fmt.Sprintf("set @@session.%v=0", variable.TiDBRedactLog))
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int, unique key (a))")
-	tk.MustExec("begin")
-	tk.MustExec("insert into t values (1),(2)")
-	c.Assert(tk.Se.GetSessionVars().PrevStmt.String(), Equals, "insert into `t` values ( ? ) , ( ? )")
-	c.Assert(tk.ExecToErr("insert into t values (1)").Error(), Equals, `[kv:1062]Duplicate entry '?' for key 'a'`)
-}
-
-func (s *testSuite) TestIssue19372(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test;")
-	tk.MustExec("drop table if exists t1, t2;")
-	tk.MustExec("create table t1 (c_int int, c_str varchar(40), key(c_str));")
-	tk.MustExec("create table t2 like t1;")
-	tk.MustExec("insert into t1 values (1, 'a'), (2, 'b'), (3, 'c');")
-	tk.MustExec("insert into t2 select * from t1;")
-	tk.MustQuery("select (select t2.c_str from t2 where t2.c_str <= t1.c_str and t2.c_int in (1, 2) order by t2.c_str limit 1) x from t1 order by c_int;").Check(testkit.Rows("a", "a", "a"))
-}
-
 func (s *testSuite) TestCollectDMLRuntimeStats(c *C) {
 	tk := testkit.NewTestKit(c, s.store)
 	tk.MustExec("use test")
@@ -2267,283 +2221,6 @@ func (s *testSuite) TestIssue20237(c *C) {
 	tk.MustExec(`insert into t values(NULL,-37), ("2011-11-04",105), ("2013-03-02",-22), ("2006-07-02",-56), (NULL,124), (NULL,111), ("2018-03-03",-5);`)
 	tk.MustExec(`insert into s values(-37),(105),(-22),(-56),(124),(105),(111),(-5);`)
 	tk.MustQuery(`select count(distinct t.a, t.b) from t join s on t.b= s.b;`).Check(testkit.Rows("4"))
-}
-
-func (s *testSerialSuite) TestIssue19148(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a decimal(16, 2));")
-	tk.MustExec("select * from t where a > any_value(a);")
-	ctx := tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(ctx).InfoSchema()
-	tblInfo, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("t"))
-	c.Assert(err, IsNil)
-	c.Assert(int(tblInfo.Meta().Columns[0].Flag), Equals, 0)
-}
-
-func (s *testSuite) TestIssue19667(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("CREATE TABLE t (a DATETIME)")
-	tk.MustExec("INSERT INTO t VALUES('1988-04-17 01:59:59')")
-	tk.MustQuery(`SELECT DATE_ADD(a, INTERVAL 1 SECOND) FROM t`).Check(testkit.Rows("1988-04-17 02:00:00"))
-}
-
-func (s *testSuite) TestZeroDateTimeCompatibility(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-
-	SQLs := []string{
-		`select YEAR(0000-00-00), YEAR("0000-00-00")`,
-		`select MONTH(0000-00-00), MONTH("0000-00-00")`,
-		`select DAYOFMONTH(0000-00-00), DAYOFMONTH("0000-00-00")`,
-		`select QUARTER(0000-00-00), QUARTER("0000-00-00")`,
-		`select EXTRACT(DAY FROM 0000-00-00), EXTRACT(DAY FROM "0000-00-00")`,
-		`select EXTRACT(MONTH FROM 0000-00-00), EXTRACT(MONTH FROM "0000-00-00")`,
-		`select EXTRACT(YEAR FROM 0000-00-00), EXTRACT(YEAR FROM "0000-00-00")`,
-		`select EXTRACT(WEEK FROM 0000-00-00), EXTRACT(WEEK FROM "0000-00-00")`,
-		`select EXTRACT(QUARTER FROM 0000-00-00), EXTRACT(QUARTER FROM "0000-00-00")`,
-	}
-	for _, t := range SQLs {
-		tk.MustQuery(t).Check(testkit.Rows("0 <nil>"))
-		c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(1))
-	}
-
-	SQLs = []string{
-		`select DAYOFWEEK(0000-00-00), DAYOFWEEK("0000-00-00")`,
-		`select DAYOFYEAR(0000-00-00), DAYOFYEAR("0000-00-00")`,
-	}
-	for _, t := range SQLs {
-		tk.MustQuery(t).Check(testkit.Rows("<nil> <nil>"))
-		c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(2))
-	}
-
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(v1 datetime, v2 datetime(3))")
-	tk.MustExec("insert ignore into t values(0,0)")
-
-	SQLs = []string{
-		`select YEAR(v1), YEAR(v2) from t`,
-		`select MONTH(v1), MONTH(v2) from t`,
-		`select DAYOFMONTH(v1), DAYOFMONTH(v2) from t`,
-		`select QUARTER(v1), QUARTER(v2) from t`,
-		`select EXTRACT(DAY FROM v1), EXTRACT(DAY FROM v2) from t`,
-		`select EXTRACT(MONTH FROM v1), EXTRACT(MONTH FROM v2) from t`,
-		`select EXTRACT(YEAR FROM v1), EXTRACT(YEAR FROM v2) from t`,
-		`select EXTRACT(WEEK FROM v1), EXTRACT(WEEK FROM v2) from t`,
-		`select EXTRACT(QUARTER FROM v1), EXTRACT(QUARTER FROM v2) from t`,
-	}
-	for _, t := range SQLs {
-		tk.MustQuery(t).Check(testkit.Rows("0 0"))
-		c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(0))
-	}
-
-	SQLs = []string{
-		`select DAYOFWEEK(v1), DAYOFWEEK(v2) from t`,
-		`select DAYOFYEAR(v1), DAYOFYEAR(v2) from t`,
-	}
-	for _, t := range SQLs {
-		tk.MustQuery(t).Check(testkit.Rows("<nil> <nil>"))
-		c.Assert(tk.Se.GetSessionVars().StmtCtx.WarningCount(), Equals, uint16(2))
-	}
-}
-
-// https://github.com/pingcap/tidb/issues/24165.
-func (s *testSuite) TestInvalidDateValueInCreateTable(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test;")
-	tk.MustExec("drop table if exists t;")
-
-	// Test for sql mode 'NO_ZERO_IN_DATE'.
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES,NO_ZERO_IN_DATE';")
-	tk.MustGetErrCode("create table t (a datetime default '2999-00-00 00:00:00');", errno.ErrInvalidDefault)
-	tk.MustExec("create table t (a datetime);")
-	tk.MustGetErrCode("alter table t modify column a datetime default '2999-00-00 00:00:00';", errno.ErrInvalidDefault)
-	tk.MustExec("drop table if exists t;")
-
-	// Test for sql mode 'NO_ZERO_DATE'.
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES,NO_ZERO_DATE';")
-	tk.MustGetErrCode("create table t (a datetime default '0000-00-00 00:00:00');", errno.ErrInvalidDefault)
-	tk.MustExec("create table t (a datetime);")
-	tk.MustGetErrCode("alter table t modify column a datetime default '0000-00-00 00:00:00';", errno.ErrInvalidDefault)
-	tk.MustExec("drop table if exists t;")
-
-	// Remove NO_ZERO_DATE and NO_ZERO_IN_DATE.
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES';")
-	// Test create table with zero datetime as a default value.
-	tk.MustExec("create table t (a datetime default '2999-00-00 00:00:00');")
-	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t (a datetime default '0000-00-00 00:00:00');")
-	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t (a datetime);")
-	tk.MustExec("alter table t modify column a datetime default '2999-00-00 00:00:00';")
-	tk.MustExec("alter table t modify column a datetime default '0000-00-00 00:00:00';")
-	tk.MustExec("drop table if exists t;")
-
-	// Test create table with invalid datetime(02-30) as a default value.
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES';")
-	tk.MustGetErrCode("create table t (a datetime default '2999-02-30 00:00:00');", errno.ErrInvalidDefault)
-	tk.MustExec("drop table if exists t;")
-	// NO_ZERO_IN_DATE and NO_ZERO_DATE have nothing to do with invalid datetime(02-30).
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES,NO_ZERO_IN_DATE,NO_ZERO_DATE';")
-	tk.MustGetErrCode("create table t (a datetime default '2999-02-30 00:00:00');", errno.ErrInvalidDefault)
-	tk.MustExec("drop table if exists t;")
-	// ALLOW_INVALID_DATES allows invalid datetime(02-30).
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES,ALLOW_INVALID_DATES';")
-	tk.MustExec("create table t (a datetime default '2999-02-30 00:00:00');")
-	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t (a datetime);")
-	tk.MustExec("alter table t modify column a datetime default '2999-02-30 00:00:00';")
-	tk.MustExec("drop table if exists t;")
-}
-
-func (s *testSuite) TestOOMActionPriority(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t0")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("drop table if exists t2")
-	tk.MustExec("drop table if exists t3")
-	tk.MustExec("drop table if exists t4")
-	tk.MustExec("create table t0(a int)")
-	tk.MustExec("insert into t0 values(1)")
-	tk.MustExec("create table t1(a int)")
-	tk.MustExec("insert into t1 values(1)")
-	tk.MustExec("create table t2(a int)")
-	tk.MustExec("insert into t2 values(1)")
-	tk.MustExec("create table t3(a int)")
-	tk.MustExec("insert into t3 values(1)")
-	tk.MustExec("create table t4(a int)")
-	tk.MustExec("insert into t4 values(1)")
-	tk.MustQuery("select * from t0 join t1 join t2 join t3 join t4 order by t0.a").Check(testkit.Rows("1 1 1 1 1"))
-	action := tk.Se.GetSessionVars().StmtCtx.MemTracker.GetFallbackForTest()
-	// check the first 5 actions is rate limit.
-	for i := 0; i < 5; i++ {
-		c.Assert(action.GetPriority(), Equals, int64(memory.DefRateLimitPriority))
-		action = action.GetFallback()
-	}
-	for action.GetFallback() != nil {
-		c.Assert(action.GetPriority(), Equals, int64(memory.DefSpillPriority))
-		action = action.GetFallback()
-	}
-	c.Assert(action.GetPriority(), Equals, int64(memory.DefLogPriority))
-}
-
-func (s testSuite) TestTrackAggMemoryUsage(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-
-	tk.MustExec("use test")
-	tk.MustExec("create table t(a int)")
-	tk.MustExec("insert into t values(1)")
-
-	tk.MustExec("set tidb_track_aggregate_memory_usage = off;")
-	rows := tk.MustQuery("explain analyze select /*+ HASH_AGG() */ sum(a) from t").Rows()
-	c.Assert(rows[0][7], Equals, "N/A")
-	rows = tk.MustQuery("explain analyze select /*+ STREAM_AGG() */ sum(a) from t").Rows()
-	c.Assert(rows[0][7], Equals, "N/A")
-	tk.MustExec("set tidb_track_aggregate_memory_usage = on;")
-	rows = tk.MustQuery("explain analyze select /*+ HASH_AGG() */ sum(a) from t").Rows()
-	c.Assert(rows[0][7], Not(Equals), "N/A")
-	rows = tk.MustQuery("explain analyze select /*+ STREAM_AGG() */ sum(a) from t").Rows()
-	c.Assert(rows[0][7], Not(Equals), "N/A")
-}
-
-func (s *testSuiteP2) TestProjectionBitType(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t(k1 int, v bit(34) DEFAULT b'111010101111001001100111101111111', primary key(k1) clustered);")
-	tk.MustExec("create table t1(k1 int, v bit(34) DEFAULT b'111010101111001001100111101111111', primary key(k1) nonclustered);")
-	tk.MustExec("insert into t(k1) select 1;")
-	tk.MustExec("insert into t1(k1) select 1;")
-
-	tk.MustExec("set @@tidb_enable_vectorized_expression = 0;")
-	// following SQL should returns same result
-	tk.MustQuery("(select * from t where false) union(select * from t for update);").Check(testkit.Rows("1 \x01\xd5\xe4\xcf\u007f"))
-	tk.MustQuery("(select * from t1 where false) union(select * from t1 for update);").Check(testkit.Rows("1 \x01\xd5\xe4\xcf\u007f"))
-
-	tk.MustExec("set @@tidb_enable_vectorized_expression = 1;")
-	tk.MustQuery("(select * from t where false) union(select * from t for update);").Check(testkit.Rows("1 \x01\xd5\xe4\xcf\u007f"))
-	tk.MustQuery("(select * from t1 where false) union(select * from t1 for update);").Check(testkit.Rows("1 \x01\xd5\xe4\xcf\u007f"))
-}
-
-func (s testSerialSuite) TestExprBlackListForEnum(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t(a enum('a','b','c'), b enum('a','b','c'), c int, index idx(b,a));")
-	tk.MustExec("insert into t values(1,1,1),(2,2,2),(3,3,3);")
-
-	checkFuncPushDown := func(rows [][]interface{}, keyWord string) bool {
-		for _, line := range rows {
-			// Agg/Expr push down
-			if line[2].(string) == "cop[tikv]" && strings.Contains(line[4].(string), keyWord) {
-				return true
-			}
-			// access index
-			if line[2].(string) == "cop[tikv]" && strings.Contains(line[3].(string), keyWord) {
-				return true
-			}
-		}
-		return false
-	}
-
-	// Test agg(enum) push down
-	tk.MustExec("insert into mysql.expr_pushdown_blacklist(name) values('enum');")
-	tk.MustExec("admin reload expr_pushdown_blacklist;")
-	rows := tk.MustQuery("desc format='brief' select /*+ HASH_AGG() */ max(a) from t;").Rows()
-	c.Assert(checkFuncPushDown(rows, "max"), IsFalse)
-	rows = tk.MustQuery("desc format='brief' select /*+ STREAM_AGG() */ max(a) from t;").Rows()
-	c.Assert(checkFuncPushDown(rows, "max"), IsFalse)
-
-	tk.MustExec("delete from mysql.expr_pushdown_blacklist;")
-	tk.MustExec("admin reload expr_pushdown_blacklist;")
-	rows = tk.MustQuery("desc format='brief' select /*+ HASH_AGG() */ max(a) from t;").Rows()
-	c.Assert(checkFuncPushDown(rows, "max"), IsTrue)
-	rows = tk.MustQuery("desc format='brief' select /*+ STREAM_AGG() */ max(a) from t;").Rows()
-	c.Assert(checkFuncPushDown(rows, "max"), IsTrue)
-
-	// Test expr(enum) push down
-	tk.MustExec("insert into mysql.expr_pushdown_blacklist(name) values('enum');")
-	tk.MustExec("admin reload expr_pushdown_blacklist;")
-	rows = tk.MustQuery("desc format='brief' select * from t where a + b;").Rows()
-	c.Assert(checkFuncPushDown(rows, "plus"), IsFalse)
-	rows = tk.MustQuery("desc format='brief' select * from t where a + b;").Rows()
-	c.Assert(checkFuncPushDown(rows, "plus"), IsFalse)
-
-	tk.MustExec("delete from mysql.expr_pushdown_blacklist;")
-	tk.MustExec("admin reload expr_pushdown_blacklist;")
-	rows = tk.MustQuery("desc format='brief' select * from t where a + b;").Rows()
-	c.Assert(checkFuncPushDown(rows, "plus"), IsTrue)
-	rows = tk.MustQuery("desc format='brief' select * from t where a + b;").Rows()
-	c.Assert(checkFuncPushDown(rows, "plus"), IsTrue)
-
-	// Test enum index
-	tk.MustExec("insert into mysql.expr_pushdown_blacklist(name) values('enum');")
-	tk.MustExec("admin reload expr_pushdown_blacklist;")
-	rows = tk.MustQuery("desc format='brief' select * from t where b = 1;").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b)"), IsFalse)
-	rows = tk.MustQuery("desc format='brief' select * from t where b = 'a';").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b)"), IsFalse)
-	rows = tk.MustQuery("desc format='brief' select * from t where b > 1;").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b)"), IsFalse)
-	rows = tk.MustQuery("desc format='brief' select * from t where b > 'a';").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b)"), IsFalse)
-
-	tk.MustExec("delete from mysql.expr_pushdown_blacklist;")
-	tk.MustExec("admin reload expr_pushdown_blacklist;")
-	rows = tk.MustQuery("desc format='brief' select * from t where b = 1 and a = 1;").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b, a)"), IsTrue)
-	rows = tk.MustQuery("desc format='brief' select * from t where b = 'a' and a = 'a';").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b, a)"), IsTrue)
-	rows = tk.MustQuery("desc format='brief' select * from t where b = 1 and a > 1;").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b, a)"), IsTrue)
-	rows = tk.MustQuery("desc format='brief' select * from t where b = 1 and a > 'a'").Rows()
-	c.Assert(checkFuncPushDown(rows, "index:idx(b, a)"), IsTrue)
 }
 
 func (s *testSuite) TestIssue24933(c *C) {
@@ -2731,195 +2408,4 @@ func (s *testSuite) TestGetResultRowsCount(c *C) {
 		cnt := executor.GetResultRowsCount(tk.Se, p)
 		c.Assert(ca.row, Equals, cnt, Commentf("sql: %v", ca.sql))
 	}
-}
-
-func checkFileName(s string) bool {
-	files := []string{
-		"config.toml",
-		"meta.txt",
-		"stats/test.t_dump_single.json",
-		"schema/test.t_dump_single.schema.txt",
-		"variables.toml",
-		"sqls.sql",
-		"session_bindings.sql",
-		"global_bindings.sql",
-		"explain.txt",
-	}
-	for _, f := range files {
-		if strings.Compare(f, s) == 0 {
-			return true
-		}
-	}
-	return false
-}
-
-// Test invoke Close without invoking Open before for each operators.
-func (s *testSerialSuite) TestUnreasonablyClose(c *C) {
-	defer testleak.AfterTest(c)()
-
-	is := infoschema.MockInfoSchema([]*model.TableInfo{plannercore.MockSignedTable(), plannercore.MockUnsignedTable()})
-	se, err := session.CreateSession4Test(s.store)
-	c.Assert(err, IsNil)
-	_, err = se.Execute(context.Background(), "use test")
-	c.Assert(err, IsNil)
-	// To enable the shuffleExec operator.
-	_, err = se.Execute(context.Background(), "set @@tidb_merge_join_concurrency=4")
-	c.Assert(err, IsNil)
-
-	var opsNeedsCovered = []plannercore.PhysicalPlan{
-		&plannercore.PhysicalHashJoin{},
-		&plannercore.PhysicalMergeJoin{},
-		&plannercore.PhysicalIndexJoin{},
-		&plannercore.PhysicalIndexHashJoin{},
-		&plannercore.PhysicalTableReader{},
-		&plannercore.PhysicalIndexReader{},
-		&plannercore.PhysicalIndexLookUpReader{},
-		&plannercore.PhysicalIndexMergeReader{},
-		&plannercore.PhysicalApply{},
-		&plannercore.PhysicalHashAgg{},
-		&plannercore.PhysicalStreamAgg{},
-		&plannercore.PhysicalLimit{},
-		&plannercore.PhysicalSort{},
-		&plannercore.PhysicalTopN{},
-		&plannercore.PhysicalCTE{},
-		&plannercore.PhysicalCTETable{},
-		&plannercore.PhysicalMaxOneRow{},
-		&plannercore.PhysicalProjection{},
-		&plannercore.PhysicalSelection{},
-		&plannercore.PhysicalTableDual{},
-		&plannercore.PhysicalWindow{},
-		&plannercore.PhysicalShuffle{},
-		&plannercore.PhysicalUnionAll{},
-	}
-	executorBuilder := executor.NewMockExecutorBuilderForTest(se, is, nil, math.MaxUint64, false, "global")
-
-	var opsNeedsCoveredMask uint64 = 1<<len(opsNeedsCovered) - 1
-	opsAlreadyCoveredMask := uint64(0)
-	for i, tc := range []string{
-		"select /*+ hash_join(t1)*/ * from t t1 join t t2 on t1.a = t2.a",
-		"select /*+ merge_join(t1)*/ * from t t1 join t t2 on t1.f = t2.f",
-		"select t.f from t use index(f)",
-		"select /*+ inl_join(t1) */ * from t t1 join t t2 on t1.f=t2.f",
-		"select /*+ inl_hash_join(t1) */ * from t t1 join t t2 on t1.f=t2.f",
-		"SELECT count(1) FROM (SELECT (SELECT min(a) FROM t as t2 WHERE t2.a > t1.a) AS a from t as t1) t",
-		"select /*+ hash_agg() */ count(f) from t group by a",
-		"select /*+ stream_agg() */ count(f) from t group by a",
-		"select * from t order by a, f",
-		"select * from t order by a, f limit 1",
-		"select * from t limit 1",
-		"select (select t1.a from t t1 where t1.a > t2.a) as a from t t2;",
-		"select a + 1 from t",
-		"select count(*) a from t having a > 1",
-		"select * from t where a = 1.1",
-		"with recursive cte1(c1) as (select 1 union select c1 + 1 from cte1 limit 5 offset 0) select * from cte1",
-		"select /*+use_index_merge(t, c_d_e, f)*/ * from t where c < 1 or f > 2",
-		"select sum(f) over (partition by f) from t",
-		"select /*+ merge_join(t1)*/ * from t t1 join t t2 on t1.d = t2.d",
-		"select a from t union all select a from t",
-	} {
-		comment := Commentf("case:%v sql:%s", i, tc)
-		c.Assert(err, IsNil, comment)
-		stmt, err := s.ParseOneStmt(tc, "", "")
-		c.Assert(err, IsNil, comment)
-
-		err = se.NewTxn(context.Background())
-		c.Assert(err, IsNil, comment)
-		p, _, err := planner.Optimize(context.TODO(), se, stmt, is)
-		c.Assert(err, IsNil, comment)
-		// This for loop level traverses the plan tree to get which operators are covered.
-		for child := []plannercore.PhysicalPlan{p.(plannercore.PhysicalPlan)}; len(child) != 0; {
-			newChild := make([]plannercore.PhysicalPlan, 0, len(child))
-			for _, ch := range child {
-				found := false
-				for k, t := range opsNeedsCovered {
-					if reflect.TypeOf(t) == reflect.TypeOf(ch) {
-						opsAlreadyCoveredMask |= 1 << k
-						found = true
-						break
-					}
-				}
-				c.Assert(found, IsTrue, Commentf("case: %v sql: %s operator %v is not registered in opsNeedsCoveredMask", i, tc, reflect.TypeOf(ch)))
-				switch x := ch.(type) {
-				case *plannercore.PhysicalCTE:
-					newChild = append(newChild, x.RecurPlan)
-					newChild = append(newChild, x.SeedPlan)
-					continue
-				case *plannercore.PhysicalShuffle:
-					newChild = append(newChild, x.DataSources...)
-					newChild = append(newChild, x.Tails...)
-					continue
-				}
-				newChild = append(newChild, ch.Children()...)
-			}
-			child = newChild
-		}
-
-		e := executorBuilder.Build(p)
-
-		func() {
-			defer func() {
-				r := recover()
-				buf := make([]byte, 4096)
-				stackSize := runtime.Stack(buf, false)
-				buf = buf[:stackSize]
-				c.Assert(r, IsNil, Commentf("case: %v\n sql: %s\n error stack: %v", i, tc, string(buf)))
-			}()
-			c.Assert(e.Close(), IsNil, comment)
-		}()
-	}
-	// The following code is used to make sure all the operators registered
-	// in opsNeedsCoveredMask are covered.
-	commentBuf := strings.Builder{}
-	if opsAlreadyCoveredMask != opsNeedsCoveredMask {
-		for i := range opsNeedsCovered {
-			if opsAlreadyCoveredMask&(1<<i) != 1<<i {
-				commentBuf.WriteString(fmt.Sprintf(" %v", reflect.TypeOf(opsNeedsCovered[i])))
-			}
-		}
-	}
-	c.Assert(opsAlreadyCoveredMask, Equals, opsNeedsCoveredMask, Commentf("these operators are not covered %s", commentBuf.String()))
-}
-
-func (s *testSerialSuite) TestEncodingSet(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("CREATE TABLE `enum-set` (`set` SET(" +
-		"'x00','x01','x02','x03','x04','x05','x06','x07','x08','x09','x10','x11','x12','x13','x14','x15'," +
-		"'x16','x17','x18','x19','x20','x21','x22','x23','x24','x25','x26','x27','x28','x29','x30','x31'," +
-		"'x32','x33','x34','x35','x36','x37','x38','x39','x40','x41','x42','x43','x44','x45','x46','x47'," +
-		"'x48','x49','x50','x51','x52','x53','x54','x55','x56','x57','x58','x59','x60','x61','x62','x63'" +
-		")NOT NULL PRIMARY KEY)")
-	tk.MustExec("INSERT INTO `enum-set` VALUES\n(\"x00,x59\");")
-	tk.MustQuery("select `set` from `enum-set` use index(PRIMARY)").Check(testkit.Rows("x00,x59"))
-	tk.MustExec("admin check table `enum-set`")
-}
-
-func (s *testSuite) TestDeleteWithMulTbl(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-
-	// Delete multiple tables from left joined table.
-	// The result of left join is (3, null, null).
-	// Because rows in t2 are not matched, so no row will be deleted in t2.
-	// But row in t1 is matched, so it should be deleted.
-	tk.MustExec("use test;")
-	tk.MustExec("drop table if exists t1, t2;")
-	tk.MustExec("create table t1 (c1 int);")
-	tk.MustExec("create table t2 (c1 int primary key, c2 int);")
-	tk.MustExec("insert into t1 values(3);")
-	tk.MustExec("insert into t2 values(2, 2);")
-	tk.MustExec("insert into t2 values(0, 0);")
-	tk.MustExec("delete from t1, t2 using t1 left join t2 on t1.c1 = t2.c2;")
-	tk.MustQuery("select * from t1 order by c1;").Check(testkit.Rows())
-	tk.MustQuery("select * from t2 order by c1;").Check(testkit.Rows("0 0", "2 2"))
-
-	// Rows in both t1 and t2 are matched, so will be deleted even if it's null.
-	// NOTE: The null values are not generated by join.
-	tk.MustExec("drop table if exists t1, t2;")
-	tk.MustExec("create table t1 (c1 int);")
-	tk.MustExec("create table t2 (c2 int);")
-	tk.MustExec("insert into t1 values(null);")
-	tk.MustExec("insert into t2 values(null);")
-	tk.MustExec("delete from t1, t2 using t1 join t2 where t1.c1 is null;")
-	tk.MustQuery("select * from t1;").Check(testkit.Rows())
-	tk.MustQuery("select * from t2;").Check(testkit.Rows())
 }
